+++ packages/twake-mui/src/index.ts
@@ -13,3 +13,6 @@ export { TwakeMuiThemeProvider } from "./components/ThemeProvider";
 export { theme } from "./lib/theme";
 export type { TwakeTheme } from "./lib/theme";
 export { radius } from "./lib/radius";
+export { Avatar, default as AvatarDefault } from "./components/Avatar";
+export type { AvatarProps } from "./components/Avatar";
+export { nameToColor, supportedColors } from "./components/Avatar/helpers";