+++ packages/twake-mui/src/lib/overrides/buttonOverrides.ts
@@ -1,5 +1,9 @@
-import { Theme, Components } from "@mui/material/styles";
+import { Theme, Components, alpha } from "@mui/material/styles";
 import { radius } from "../radius";
+import paletteJson from "../palette.json";
+import { PaletteJson } from "../types";
+
+const paletteData = paletteJson as PaletteJson;
 
 export const buttonOverrides = (theme: Theme): Components["MuiButton"] => {
   return {
@@ -29,8 +33,32 @@ export const buttonOverrides = (theme: Theme): Components["MuiButton"] => {
           boxShadow: "none",
         },
       },
+      containedSecondary: {
+        backgroundColor: theme.palette.secondary.main,
+        color: alpha(theme.palette.text.primary, 0.9),
+        borderRadius: radius.sm,
+        "&:hover": {
+          backgroundColor: paletteData.Secondary[700],
+          color: alpha(theme.palette.text.primary, 0.9),
+        },
+        "&.Mui-disabled": {
+          backgroundColor: theme.palette.action.disabledBackground,
+        },
+      },
       outlined: {
-        // Outlined button overrides
+        backgroundColor: "#fff",
+        borderColor: alpha(paletteData.Grey.A900, 0.28),
+        color: alpha(theme.palette.text.primary, 0.9),
+        "&:hover": {
+          backgroundColor: alpha(theme.palette.primary.main, 0.04),
+          color: theme.palette.primary.main,
+          borderColor: alpha(theme.palette.primary.main, 0.32),
+        },
+        "&.Mui-disabled": {
+          backgroundColor: "#fff",
+          color: alpha(theme.palette.grey[900], 0.38),
+          borderColor: alpha(paletteData.Grey.A900, 0.28),
+        },
       },
       text: {
         // Text button overrides